requirements.txt
openai
typer
pyyaml
jinja2
